15-11-2022
- Validate Enum Attribute
- Change name service interfaces
- Repository Layer
- Generic Repository
