+++ tlcn_dotnet/Controllers/CategoryController.cs
@@ -13,8 +13,8 @@ namespace tlcn_dotnet.Controllers
     [Route("[controller]")]
     public class CategoryController : ControllerBase
     {
-        private readonly CategoryService _categoryService;
-        public CategoryController(CategoryService categoryService)
+        private readonly ICategoryService _categoryService;
+        public CategoryController(ICategoryService categoryService)
         {
             _categoryService = categoryService;
         }
