+++ tlcn_dotnet/Controllers/ProductController.cs
@@ -15,8 +15,8 @@ namespace tlcn_dotnet.Controllers
     [ApiController]
     public class ProductController : ControllerBase
     {
-        private readonly ProductService _productService;
-        public ProductController(ProductService productService)
+        private readonly IProductService _productService;
+        public ProductController(IProductService productService)
         {
             _productService = productService;
         }
