+++ tlcn_dotnet/IRepositories/ICategoryRepository.cs
@@ -1,0 +1,9 @@
+﻿using tlcn_dotnet.Entity;
+
+namespace tlcn_dotnet.IRepositories
+{
+    public interface ICategoryRepository: IGenericRepository<Category>
+    {
+        
+    }
+}
