+++ tlcn_dotnet/IRepositories/ISupplierRepository.cs
@@ -1,0 +1,8 @@
+﻿using tlcn_dotnet.Entity;
+
+namespace tlcn_dotnet.IRepositories
+{
+    public interface ISupplierRepository: IGenericRepository<Supplier>
+    {
+    }
+}
