+++ tlcn_dotnet/IServices/IConfirmTokenService.cs
@@ -1,0 +1,9 @@
+﻿using tlcn_dotnet.Entity;
+
+namespace tlcn_dotnet.Services
+{
+    public interface IConfirmTokenService
+    {
+        public Task<ConfirmToken> CreateConfirmToken(Account account);
+    }
+}
