+++ tlcn_dotnet/IServices/IEmailService.cs
@@ -1,0 +1,10 @@
+﻿using MimeKit;
+using tlcn_dotnet.Entity;
+
+namespace tlcn_dotnet.Services
+{
+    public interface IEmailService
+    {
+        public Task SendRegisterConfirmationToken(ConfirmToken confirmToken);
+    }
+}
