+++ tlcn_dotnet/IServices/ILocationService.cs
@@ -1,0 +1,8 @@
+﻿namespace tlcn_dotnet.Services
+{
+    public interface ILocationService
+    {
+        public Task<DataResponse> GetAllCountryAndCity();
+        public Task<DataResponse> GetVietnamLocation();
+    }
+}
